projects/aurora-oceano/requirements.txt
# Aurora Oceano - Dependencies
# Python 3.10+

# Core
requests>=2.31.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Telegram (opcional - se usar python-telegram-bot)
# python-telegram-bot>=20.0

# PDF Generation
# reportlab>=4.0.0
# fpdf2>=2.7.0
